+++ CHANGELOG.md
@@ -9,6 +9,7 @@ To see compatibility information, `xsvutils help compatibility`.
 - add some subcommands on v2
 - add --src, --dst options of inshour, insdate subcommand on v2
 - rename subcommand name `insdeltasec` to `inssecinterval`
+- remove `addlinenum2` subcommand
 
 ## version 0.2.2 (2018-03-07)
 - add `inshour` subcommand on v2